backend/requirements.txt
flask
flask-cors
psycopg2-binary
numpy
scipy
scikit-learn
joblib
bcrypt
PyJWT
pandas
python-dotenv